safari-600.7-branch/LayoutTests/editing/deleting/smart-delete-002.html

<html> 
<head>

<style>
.editing { 
    border: 2px solid red; 
    font-size: 24px; 
}
.explanation { 
    border: 2px solid blue; 
    padding: 12px; 
    font-size: 24px; 
    margin-bottom: 24px;
}
.scenario { margin-bottom: 16px;}
.scenario:first-line { font-weight: bold; margin-bottom: 16px;}
.expected-results:first-line { font-weight: bold }
</style>
<script src=../editing.js language="JavaScript" type="text/JavaScript" ></script>

<script>

function editingTest() {
    if (window.internals)
        internals.settings.setEditingBehavior('mac');
    moveSelectionForwardByWordCommand();
    moveSelectionForwardByCharacterCommand();
    doubleClickAtSelectionStart();
    deleteCommand();
}

</script>

<title>Editing Test</title> 
</head> 
<body>

<div class="explanation">
<div class="scenario">
Tests: 
<br>
Smart delete when deleting the second word of a line.
</div>
<div class="expected-results">
Expected Results:
<br>
The second word and the space before the second word should be deleted. It should like this this:
<BR>
foo baz
</div>
</div>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div id="test" class="editing">
foo bar baz
</div>
</div>

<script>
runEditingTest();
</script>

</body>
</html>
